blogtwo/requirements.txt
absl-py==0.7.1
aiodns==1.1.1
aiohttp==3.2.0
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
asn1crypto==0.24.0
astor==0.8.0
astroid==2.2.5
async-timeout==3.0.0
attrs==19.1.0
Automat==0.6.0
backcall==0.1.0
beautifulsoup4==4.6.0
billiard==3.3.0.23
bleach==3.1.0
blinker==1.4
bottle==0.12.17
brotlipy==0.7.0
cchardet==2.1.1
celery==3.1.26.post2
celery-with-redis==3.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
click==6.7
colorama==0.4.1
constantly==15.1.0
coverage==4.5.4
cryptography==2.7
cssselect==1.0.3
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
Django==1.11.9
django-appconf==1.0.3
django-autoslug==1.9.6
django-celery==3.2.2
django-compressor==2.3
django-debug-toolbar==2.0
django-haystack==2.8.1
django-ipware==2.1.0
django-mdeditor==0.1.16
django-pure-pagination==0.3.0
django-redis==4.9.0
django-redis-sessions==0.5.6
django-tinymce==2.7.0
django-uuslug==1.1.8
djangorestframework==3.8.2
elasticsearch==7.0.0
elasticsearch-dsl==7.0.0
entrypoints==0.3
Pillow==6.1.0
pymongo==3.6.1
PyMySQL==0.8.1
python-docx==0.8.10
pytz==2019.2
pywin32==224
redis==2.10.6
requests==2.22.0
retrying==1.3.3
rjsmin==1.1.0
ruamel.yaml==0.15.44
scikit-learn==0.21.2
scipy==1.2.1
Scrapy==1.6.0
scrapy-redis==0.6.8
selenium==3.12.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.12.0
sortedcontainers==2.0.4
SQLAlchemy==1.3.3
sqlparse==0.3.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.2
tesserocr==2.2.2
testpath==0.4.2
text-unidecode==1.2
tornado==5.0.2
traitlets==4.3.2
Twisted==18.4.0
typed-ast==1.3.1
Unidecode==1.1.1
urllib3==1.25.3
urwid==2.0.1
vboxapi==1.0
vine==1.1.4
virtualenv==15.1.0
w3lib==1.20.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
WeRoBot==1.9.0
Whoosh==2.7.4
widgetsnbextension==3.4.2
wrapt==1.11.2
wsproto==0.11.0
xmltodict==0.12.0
yarl==1.2.4
zope.interface==4.5.0
